# src/agent/Hercules.WebApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Hercules.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/simulation/templates/{templateName}/available": {
      "get": {
        "tags": [
          "Simulation"
        ],
        "operationId": "SimulationAvailability",
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/simulation/sessions": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "operationId": "StartSimulationSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/simulation/sessions/{templateName}/replay": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "operationId": "SimulationReplay",
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/simulation/sessions/{templateName}": {
      "delete": {
        "tags": [
          "Simulation"
        ],
        "operationId": "StopSimulationSession",
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Simulation"
        ],
        "operationId": "GetSimulationSession",
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/maintenance/run": {
      "post": {
        "tags": [
          "SelfImprovement"
        ],
        "operationId": "RunMaintenance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RunMaintenanceRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenance/run-all": {
      "post": {
        "tags": [
          "SelfImprovement"
        ],
        "operationId": "RunMaintenanceAll",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RunMaintenanceRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenance/proposals": {
      "get": {
        "tags": [
          "SelfImprovement"
        ],
        "operationId": "ListProposals",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "skillId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenance/proposals/{id}": {
      "get": {
        "tags": [
          "SelfImprovement"
        ],
        "operationId": "GetProposal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenance/proposals/{id}/approve": {
      "post": {
        "tags": [
          "SelfImprovement"
        ],
        "operationId": "ApproveProposal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ApproveRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenance/proposals/{id}/reject": {
      "post": {
        "tags": [
          "SelfImprovement"
        ],
        "operationId": "RejectProposal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RejectRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveRequest": {
        "type": "object",
        "properties": {
          "approvedBy": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RejectRequest": {
        "type": "object",
        "properties": {
          "rejectedBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ReplayRequest": {
        "type": "object",
        "properties": {
          "withFailures": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "RunMaintenanceRequest": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string"
          },
          "triggeredBy": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StartSessionRequest": {
        "required": [
          "templateName"
        ],
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}